% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Compute_LogP_LogB_LogY.R
\name{Compute_Yield_Biomass_Production}
\alias{Compute_Yield_Biomass_Production}
\title{Species-specific production, yield and biomass beyond a given size cutoff
this cutoff should be specified in weight and can either be a
species-specific list or a single value for all species, if a size cutoff is
not supplied the cutoff will be taken as the l50 of the gears using a
weighted average by the catchabilities.}
\usage{
Compute_Yield_Biomass_Production(
  object,
  size = NULL,
  length = NULL,
  n = NULL,
  n_pp = NULL,
  n_other = NULL,
  effort = NULL,
  t = NULL,
  ...
)
}
\description{
Species-specific production, yield and biomass beyond a given size cutoff
this cutoff should be specified in weight and can either be a
species-specific list or a single value for all species, if a size cutoff is
not supplied the cutoff will be taken as the l50 of the gears using a
weighted average by the catchabilities.
}
