% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Compute_LogP_LogB_LogY.R
\name{plot_LogB_LogP}
\alias{plot_LogB_LogP}
\alias{plot_LogB_LogP.MizerParams}
\alias{plot_LogB_LogP.MizerSim}
\title{Plot of the Biomass Production relationship}
\usage{
plot_LogB_LogP(object, ...)

\method{plot_LogB_LogP}{MizerParams}(
  object,
  size = NULL,
  length = NULL,
  n = initialN(object),
  n_pp = initialNResource(object),
  n_other = initialNOther(object),
  ...
)

\method{plot_LogB_LogP}{MizerSim}(object, size = NULL, length = NULL, ...)
}
\description{
Plot of the Biomass Production relationship
}
