% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Compute_LogP_LogB_LogY.R
\name{plot_LogY_LogP}
\alias{plot_LogY_LogP}
\alias{plot_LogY_LogP.MizerParams}
\alias{plot_LogY_LogP.MizerSim}
\title{Plot Log Y Log P relationship}
\usage{
plot_LogY_LogP(object, ...)

\method{plot_LogY_LogP}{MizerParams}(
  object,
  size = NULL,
  length = NULL,
  n = initialN(object),
  n_pp = initialNResource(object),
  n_other = initialNOther(object),
  ...
)

\method{plot_LogY_LogP}{MizerSim}(object, size = NULL, length = NULL, ...)
}
\description{
Plot Log Y Log P relationship
}
