% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Compute_LogP_LogB_LogY.R
\name{plotBiomassProductionSlope}
\alias{plotBiomassProductionSlope}
\alias{plotBiomassProductionSlope.MizerSim}
\title{Plot log10(Biomass) vs log10(Production) Slope Over Time}
\usage{
plotBiomassProductionSlope(
  object,
  sim2 = NULL,
  species = NULL,
  tlim = c(NA, NA),
  return_data = FALSE,
  size = NULL,
  length = NULL,
  ...
)

\method{plotBiomassProductionSlope}{MizerSim}(
  object,
  sim2 = NULL,
  species = NULL,
  tlim = c(NA, NA),
  return_data = FALSE,
  size = NULL,
  length = NULL,
  ...
)
}
\description{
Plot log10(Biomass) vs log10(Production) Slope Over Time
}
