% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Compute_LogP_LogB_LogY.R
\name{precompute_cutoff}
\alias{precompute_cutoff}
\title{Helper function that precomputes the weights corresponding to the desired
size cutoff for the metrics such as Yield, Biomass and Production
Precompute the params for species-selective balanced harvest}
\usage{
precompute_cutoff(object, size = NULL, length = NULL, ...)
}
\description{
This function precomputes static properties ONCE before running
simulations
}
