% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetDiffusion.R
\name{setDiffusion}
\alias{setDiffusion}
\title{Set diffusion rate}
\usage{
setDiffusion(object, diffusion = NULL, reset = FALSE, ...)
}
\arguments{
\item{object}{A MizerParams object}

\item{diffusion}{Optional. An array (species x size) holding the diffusion rate.}

\item{reset}{If set to TRUE, the diffusion rate will be reset to the
calculated value even if previously overwritten manually.}

\item{...}{Unused}
}
\value{
A MizerParams object with updated diffusion rate.
}
\description{
Set diffusion rate
}
