% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneParams.R
\name{tuneGrowth}
\alias{tuneGrowth}
\title{Tune Growth}
\usage{
tuneGrowth(params, match = "biomass")
}
\arguments{
\item{params}{MizerParams object to tune. If missing, the gadget tries to
recover information from log files left over from aborted previous runs.}

\item{match}{A character vector. Determines which quantities should be
matched to observations each time the "steady" button is pressed. Possible
entries are \code{"growth"} (using \code{\link[mizer:matchGrowth]{matchGrowth()}}), \code{"biomass"} (using
\code{\link[mizer:matchBiomasses]{matchBiomasses()}}) and \code{"yield"} (using \code{\link[=matchYields]{matchYields()}}).}
}
\value{
The tuned MizerParams object
}
\description{
A simplified instance of \code{tuneParams()} that is useful for tuning growth
rates.
}
